fix: 루나팀 신호 승인 상태 전환 누락 버그 수정
- luna.js: 네메시스 승인 시 updateSignalStatus('approved') 추가
  → 헤파이스토스가 getApprovedSignals()로 신호를 감지하지 못하던 문제 해결
  → 이제 승인된 신호가 실제 바이낸스 주문으로 이어짐
- luna.js: 네메시스 거부 시 updateSignalStatus('rejected') 추가
  → 거부 이력 추적 가능
- reporter.js: ROUND(double precision, integer) → ROUND(::numeric) 캐스트 수정

+++ bots/investment/team/reporter.js
@@ -84,7 +84,7 @@ export async function generateReport({ days = 30, telegram = false } = {}) {
     SELECT
       status,
       COUNT(*)::INTEGER  AS cnt,
-      ROUND(AVG(confidence) * 100, 1) AS avg_conf
+      ROUND(AVG(confidence)::numeric * 100, 1) AS avg_conf
     FROM signals
     WHERE created_at > now() - INTERVAL '${days} days'
     GROUP BY status
